feat(extend): on_palette can override extended palette;
`on_palette` is now called at the start and end of `extend_palette`.

**previously:**

If you set a color like `palette.bg` it would be overwritten by extend palette.

**now:**

It will be overwritten at the start and then set again at the end.
while still allowing changes to the base palette to propagate.